+++ app/Http/Controllers/Teacher/Question/Writing/Dialog/DialogController.php
@@ -7,8 +7,12 @@ use App\Http\Controllers\Controller;
 use App\Model\Writing\Dialog;
 use App\Model\Writing\WritingPart;
 use App\QuestionSet;
+use Illuminate\Contracts\View\Factory;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Crypt;
+use Illuminate\View\View;
 
 class DialogController extends Controller
 {
@@ -21,7 +25,7 @@ class DialogController extends Controller
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @return Factory|View
      */
     public function index()
     {
@@ -31,7 +35,7 @@ class DialogController extends Controller
     /**
      * Show the form for creating a new resource.
      *
-     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @return Factory|View
      */
     public function create()
     {
@@ -43,15 +47,16 @@ class DialogController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param \Illuminate\Http\Request $request
-     * @return \Illuminate\Http\RedirectResponse
+     * @param Request $request
+     * @return RedirectResponse
      */
     public function store(Request $request)
     {
         $exam = $request->exam;
         $set = $request->questionSet;
 
-        $countDialogs = Dialog::where(['exam_id' => $exam, 'question_set_id' => $set])->get()->count();
+        $authTeacher = Auth::guard('teacher')->user();
+        $countDialogs = $authTeacher->exams()->find($exam)->dialogs()->where(['question_set_id' => $set])->get()->count();
 
         if ($countDialogs < 1) {
             Dialog::create($this->validateDialogCreateRequest($request));
@@ -59,7 +64,7 @@ class DialogController extends Controller
             toast('Dialog has been successfully added','success');
         } else {
             session()->flash('field_audio');
-            alert()->info('Fail!', 'You can no longer add dialog to this set.');
+            alert()->info('Fail!', 'You can no longer add dialog to this '. QuestionSet::find($set)->name .' set.');
         }
         return redirect()->back();
     }
@@ -67,57 +72,90 @@ class DialogController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param \App\Model\Writing\Dialog $dialog
-     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @param Dialog $dialog
+     * @return Factory|RedirectResponse|View
      */
     public function show(Dialog $dialog)
     {
-        return view('teacher.questions.writing.dialogs.show', compact('dialog'))
-            ->with('authTeacherExams', Auth::guard('teacher')->user()->exams)
-            ->with('questionSets', QuestionSet::all());
+        if ($this->validDialogRequest($dialog)) {
+            return view('teacher.questions.writing.dialogs.show', compact('dialog'))
+                ->with('authTeacherExams', Auth::guard('teacher')->user()->exams)
+                ->with('questionSets', QuestionSet::all());
+        } else {
+            alert()->error('😒', 'You can\'t do this.');
+            return redirect()->back();
+        }
     }
 
     /**
      * Show the form for editing the specified resource.
      *
-     * @param \App\Model\Writing\Dialog $dialog
-     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @param Dialog $dialog
+     * @return Factory|RedirectResponse|View
      */
     public function edit(Dialog $dialog)
     {
-        return view('teacher.questions.writing.dialogs.edit', compact('dialog'))
-            ->with('authTeacherExams', Auth::guard('teacher')->user()->exams()->latest()->get())
-            ->with('questionSets', QuestionSet::all());
+        if ($this->validDialogRequest($dialog)) {
+            return view('teacher.questions.writing.dialogs.edit', compact('dialog'))
+                ->with('authTeacherExams', Auth::guard('teacher')->user()->exams()->latest()->get())
+                ->with('questionSets', QuestionSet::all());
+        } else {
+            alert()->error('😒', 'You can\'t do this.');
+            return redirect()->back();
+        }
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param \Illuminate\Http\Request $request
-     * @param \App\Model\Writing\Dialog $dialog
-     * @return \Illuminate\Http\RedirectResponse
+     * @param Request $request
+     * @param Dialog $dialog
+     * @return RedirectResponse
      */
     public function update(Request $request, Dialog $dialog)
     {
-        $dialog->update($this->validateDialogUpdateRequest($request));
-        session()->flash('success_audio');
-        toast('Dialog has been successfully updated','success');
-        return redirect()->route('teachers.questions.dialogs.show', $dialog->id);
+        if ($this->validDialogRequest($dialog)) {
+            $dialog->update($this->validateDialogUpdateRequest($request));
+            session()->flash('success_audio');
+            toast('Dialog has been successfully updated','success');
+            return redirect(route('teachers.questions.dialogs.show', $dialog->id).'?exam='.\request()->get('exam'));
+        } else {
+            alert()->error('😒', 'You can\'t do this.');
+            return redirect()->back();
+        }
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param \App\Model\Writing\Dialog $dialog
-     * @return \Illuminate\Http\RedirectResponse
+     * @param Dialog $dialog
+     * @return RedirectResponse
      */
     public function destroy(Dialog $dialog)
     {
-        $dialog->forceDelete();
+        if ($this->validDialogRequest($dialog)) {
+            $dialog->forceDelete();
+            session()->flash('success_audio');
+            toast('Dialog has been successfully deleted','success');
+            return redirect()->route('teachers.questions.dialogs.index');
+        } else {
+            alert()->error('😒', 'You can\'t do this.');
+            return redirect()->back();
+        }
+    }
 
-        session()->flash('success_audio');
-        toast('Dialog has been successfully deleted','success');
-        return redirect()->route('teachers.questions.dialogs.index');
+    private function validDialogRequest($dialog) {
+
+        $examId = Crypt::decrypt(\request()->get('exam'));
+        $authTeacherDialogs = Auth::guard('teacher')->user()->exams()->find($examId)->dialogs;
+        $valid = null;
+        foreach ($authTeacherDialogs as $authTeacherDialog) {
+            if ($authTeacherDialog->id === $dialog->id) {
+                $valid = true;
+            }
+        }
+
+        return $valid;
     }
 
     private function validateDialogCreateRequest($request)
@@ -134,7 +172,6 @@ class DialogController extends Controller
         return [
             'exam_id' => $validateData['exam'],
             'question_set_id' => $validateData['questionSet'],
-            'writing_part_id' => WritingPart::where('slug', 'dialog')->get()->first()->id,
             'topic' => $validateData['topic'],
             'question_1' => $validateData['question_1'],
             'question_2' => $validateData['question_2'],
@@ -142,6 +179,7 @@ class DialogController extends Controller
         ];
 
     }
+
     private function validateDialogUpdateRequest($request)
     {
         $validateData = $this->validate($request, [
